Remove all Settings DataStore usage
