+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/datastore/DataStoreType.kt
@@ -3,6 +3,5 @@ package com.woocommerce.android.wear.datastore
 enum class DataStoreType(val typeName: String) {
     LOGIN("login"),
     STATS("stats"),
-    ORDERS("orders"),
-    SETTINGS("settings")
+    ORDERS("orders")
 }
