+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/wear/di/DataStoreModule.kt
@@ -8,7 +8,6 @@ import androidx.datastore.preferences.preferencesDataStoreFile
 import com.woocommerce.android.wear.datastore.DataStoreQualifier
 import com.woocommerce.android.wear.datastore.DataStoreType.LOGIN
 import com.woocommerce.android.wear.datastore.DataStoreType.ORDERS
-import com.woocommerce.android.wear.datastore.DataStoreType.SETTINGS
 import com.woocommerce.android.wear.datastore.DataStoreType.STATS
 import dagger.Module
 import dagger.Provides
@@ -51,17 +50,6 @@ class DataStoreModule {
         @AppCoroutineScope appCoroutineScope: CoroutineScope
     ): DataStore<Preferences> = PreferenceDataStoreFactory.create(
         produceFile = { appContext.preferencesDataStoreFile(ORDERS.typeName) },
-        scope = CoroutineScope(appCoroutineScope.coroutineContext + Dispatchers.IO)
-    )
-
-    @Provides
-    @Singleton
-    @DataStoreQualifier(SETTINGS)
-    fun providePrefsDataStore(
-        appContext: Context,
-        @AppCoroutineScope appCoroutineScope: CoroutineScope
-    ): DataStore<Preferences> = PreferenceDataStoreFactory.create(
-        produceFile = { appContext.preferencesDataStoreFile(SETTINGS.typeName) },
         scope = CoroutineScope(appCoroutineScope.coroutineContext + Dispatchers.IO)
     )
 }
